Cleanup unused types
