+++ src/logic/safe/store/actions/__tests__/utils.test.ts
@@ -1,99 +1,158 @@
-import { getNewTxNonce, shouldExecuteTransaction } from 'src/logic/safe/store/actions/utils'
-import { GnosisSafe } from 'src/types/contracts/GnosisSafe.d'
-import { TxServiceModel } from 'src/logic/safe/store/actions/transactions/fetchTransactions/loadOutgoingTransactions'
-import { getMockedSafeInstance } from 'src/test/utils/safeHelper'
-import { NonPayableTransactionObject } from 'src/types/contracts/types'
+import { getLastTx, getNewTxNonce, shouldExecuteTransaction } from 'src/logic/safe/store/actions/utils'
+import { getMockedSafeInstance, getMockedTxServiceModel } from 'src/test/utils/safeHelper'
+import axios from 'axios'
+import { buildTxServiceUrl } from 'src/logic/safe/transactions'
 
-describe('Store actions utils > getNewTxNonce', () => {
-  it(`Should return nonce of a last transaction + 1 if passed nonce is less than last transaction or invalid`, async () => {
-    // Given
-    const lastTx = { nonce: 44 } as TxServiceModel
-    const safeInstance = {
-      methods: {
-        nonce: () => ({
-          call: () => Promise.resolve('45'),
-        }),
-      },
-    }
+describe('shouldExecuteTransaction', () => {
+  it('It should return false if given a safe with a threshold > 1', async () => {
+    // given
+    const nonce = '0'
+    const threshold = '2'
+    const safeInstance = getMockedSafeInstance({ threshold })
+    const lastTx = getMockedTxServiceModel({})
 
-    // When
-    const nonce = await getNewTxNonce(lastTx, safeInstance as GnosisSafe)
+    // when
+    const result = await shouldExecuteTransaction(safeInstance, nonce, lastTx)
 
-    // Then
-    expect(nonce).toBe('45')
+    // then
+    expect(result).toBe(false)
   })
+  it('It should return true if given a safe with a threshold === 1 and the previous transaction is already executed', async () => {
+    // given
+    const nonce = '1'
+    const threshold = '1'
+    const safeInstance = getMockedSafeInstance({ threshold, nonce })
+    const lastTx = getMockedTxServiceModel({})
 
-  it(`Should retrieve contract's instance nonce value as a fallback, if txNonce and lastTx are not valid`, async () => {
-    // Given
-    const lastTx = null
-    const safeInstance = {
-      methods: {
-        nonce: () => ({
-          call: () => Promise.resolve('45'),
-        }),
-      },
-    }
+    // when
+    const result = await shouldExecuteTransaction(safeInstance, nonce, lastTx)
 
-    // When
-    const nonce = await getNewTxNonce(lastTx, safeInstance as GnosisSafe)
+    // then
+    expect(result).toBe(true)
+  })
+  it('It should return true if given a safe with a threshold === 1 and the previous transaction is already executed', async () => {
+    // given
+    const nonce = '10'
+    const threshold = '1'
+    const safeInstance = getMockedSafeInstance({ threshold, nonce })
+    const lastTx = getMockedTxServiceModel({ isExecuted: true })
 
-    // Then
-    expect(nonce).toBe('45')
+    // when
+    const result = await shouldExecuteTransaction(safeInstance, nonce, lastTx)
+
+    // then
+    expect(result).toBe(true)
+  })
+  it('It should return false if given a safe with a threshold === 1 and the previous transaction is not yet executed', async () => {
+    // given
+    const nonce = '10'
+    const threshold = '1'
+    const safeInstance = getMockedSafeInstance({ threshold })
+    const lastTx = getMockedTxServiceModel({ isExecuted: false })
+
+    // when
+    const result = await shouldExecuteTransaction(safeInstance, nonce, lastTx)
+
+    // then
+    expect(result).toBe(false)
   })
 })
 
-describe('Store actions utils > shouldExecuteTransaction', () => {
-  it(`should return false if there's a previous tx pending to be executed`, async () => {
-    // Given
+describe('getNewTxNonce', () => {
+  it('It should return 2 if given the last transaction with nonce 1', async () => {
+    // given
     const safeInstance = getMockedSafeInstance({})
-    safeInstance.methods.getThreshold = () =>
-      ({
-        call: () => Promise.resolve('1'),
-      } as NonPayableTransactionObject<string>)
+    const lastTx = getMockedTxServiceModel({ nonce: 1 })
+    const expectedResult = '2'
 
-    const nonce = '1'
-    const lastTx = { isExecuted: false } as TxServiceModel
+    // when
+    const result = await getNewTxNonce(lastTx, safeInstance)
 
-    // When
-    const isExecution = await shouldExecuteTransaction(safeInstance as GnosisSafe, nonce, lastTx)
-
-    // Then
-    expect(isExecution).toBeFalsy()
+    // then
+    expect(result).toBe(expectedResult)
   })
+  it('It should return 0 if given a safe with nonce 0 and no transactions should use safe contract instance for retrieving nonce', async () => {
+    // given
+    const safeNonce = '0'
+    const safeInstance = getMockedSafeInstance({ nonce: safeNonce })
+    const expectedResult = '0'
+    const mockFnCall = jest.fn().mockImplementation(() => safeNonce)
+    const mockFnNonce = jest.fn().mockImplementation(() => ({ call: mockFnCall }))
 
-  it(`should return false if threshold is greater than 1`, async () => {
-    // Given
+    safeInstance.methods.nonce = mockFnNonce
+
+    // when
+    const result = await getNewTxNonce(null, safeInstance)
+
+    // then
+    expect(result).toBe(expectedResult)
+    expect(mockFnNonce).toHaveBeenCalled()
+    expect(mockFnCall).toHaveBeenCalled()
+    mockFnNonce.mockRestore()
+    mockFnCall.mockRestore()
+  })
+  it('Given a Safe and the last transaction, should return nonce of the last transaction + 1', async () => {
+    // given
     const safeInstance = getMockedSafeInstance({})
-    safeInstance.methods.getThreshold = () =>
-      ({
-        call: () => Promise.resolve('2'),
-      } as NonPayableTransactionObject<string>)
+    const expectedResult = '11'
+    const lastTx = getMockedTxServiceModel({ nonce: 10 })
 
-    const nonce = '1'
-    const lastTx = { isExecuted: true } as TxServiceModel
+    // when
+    const result = await getNewTxNonce(lastTx, safeInstance)
 
-    // When
-    const isExecution = await shouldExecuteTransaction(safeInstance as GnosisSafe, nonce, lastTx)
-
-    // Then
-    expect(isExecution).toBeFalsy()
+    // then
+    expect(result).toBe(expectedResult)
   })
+})
 
-  it(`should return true is threshold is 1 and previous tx is executed`, async () => {
-    // Given
-    const safeInstance = getMockedSafeInstance({ nonce: '1' })
-    safeInstance.methods.getThreshold = () =>
-      ({
-        call: () => Promise.resolve('1'),
-      } as NonPayableTransactionObject<string>)
+jest.mock('axios')
+jest.mock('console')
+describe('getLastTx', () => {
+  afterAll(() => {
+    jest.unmock('axios')
+    jest.unmock('console')
+  })
+  const safeAddress = '0xdfA693da0D16F5E7E78FdCBeDe8FC6eBEa44f1Cf'
+  it('It should return the last transaction for a given a safe address', async () => {
+    // given
+    const lastTx = getMockedTxServiceModel({ nonce: 1 })
+    const url = buildTxServiceUrl(safeAddress)
 
-    const nonce = '1'
-    const lastTx = { isExecuted: true } as TxServiceModel
+    // when
+    // @ts-ignore
+    axios.get.mockImplementationOnce(() => {
+      return {
+        data: {
+          results: [lastTx],
+        },
+      }
+    })
 
-    // When
-    const isExecution = await shouldExecuteTransaction(safeInstance as GnosisSafe, nonce, lastTx)
+    const result = await getLastTx(safeAddress)
 
-    // Then
-    expect(isExecution).toBeTruthy()
+    // then
+    expect(result).toStrictEqual(lastTx)
+    expect(axios.get).toHaveBeenCalled()
+    expect(axios.get).toBeCalledWith(url, { params: { limit: 1 } })
+  })
+  it('If should return null If catches an error getting last transaction', async () => {
+    // given
+    const lastTx = null
+    const url = buildTxServiceUrl(safeAddress)
+
+    // when
+    // @ts-ignore
+    axios.get.mockImplementationOnce(() => {
+      throw new Error()
+    })
+    console.error = jest.fn()
+    const result = await getLastTx(safeAddress)
+    const spyConsole = jest.spyOn(console, 'error').mockImplementation()
+
+    // then
+    expect(result).toStrictEqual(lastTx)
+    expect(axios.get).toHaveBeenCalled()
+    expect(axios.get).toBeCalledWith(url, { params: { limit: 1 } })
+    expect(spyConsole).toHaveBeenCalled()
   })
 })
