+++ src/logic/safe/store/models/types/transactions.d.ts
@@ -116,73 +116,6 @@ export interface Transfer {
   from: string
 }
 
-export enum TxType {
-  MULTISIG_TRANSACTION = 'MULTISIG_TRANSACTION',
-  ETHEREUM_TRANSACTION = 'ETHEREUM_TRANSACTION',
-  MODULE_TRANSACTION = 'MODULE_TRANSACTION',
-}
-
-export interface MultiSigTransaction {
-  safe: string
-  to: string
-  value: string
-  data: string | null
-  operation: number
-  gasToken: string
-  safeTxGas: number
-  baseGas: number
-  gasPrice: string
-  refundReceiver: string
-  nonce: number
-  executionDate: string | null
-  submissionDate: string
-  modified: string
-  blockNumber: number | null
-  transactionHash: string | null
-  safeTxHash: string
-  executor: string | null
-  isExecuted: boolean
-  isSuccessful: boolean | null
-  ethGasPrice: string | null
-  gasUsed: number | null
-  fee: string | null
-  origin: string | null
-  dataDecoded: DataDecoded | null
-  confirmationsRequired: number | null
-  confirmations: Confirmation[]
-  signatures: string | null
-  transfers: Transfer[]
-  txType: TxType.MULTISIG_TRANSACTION
-}
-
-export interface ModuleTransaction {
-  created: string
-  executionDate: string
-  blockNumber: number
-  transactionHash: string
-  safe: string
-  module: string
-  to: string
-  value: string
-  data: string
-  operation: Operation
-  transfers: Transfer[]
-  txType: TxType.MODULE_TRANSACTION
-}
-
-export interface EthereumTransaction {
-  executionDate: string
-  to: string
-  data: string | null
-  txHash: string
-  blockNumber: number
-  transfers: Transfer[]
-  txType: TxType.ETHEREUM_TRANSACTION
-  from: string
-}
-
-export type Transaction = MultiSigTransaction | ModuleTransaction | EthereumTransaction
-
 // SAFE METHODS TO ITS ID
 // https://github.com/gnosis/safe-contracts/blob/development/test/safeMethodNaming.js
 // https://github.com/gnosis/safe-contracts/blob/development/contracts/GnosisSafe.sol
